requirements-dev.txt
ruff==0.3.6
black==24.3.0

mysql-connector-python==8.2.0
